[refactor] Move ti funcs out of taichi.lang.
